refactor(pageserver): check layer map valid in one place
